RAG/requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
faiss-cpu==1.7.4
wikipedia==1.4.0
python-dotenv==1.0.1
deep-translator==1.11.4 